refactor: replace debug functions
