+++ src/Element/Plane/plane.h
@@ -30,8 +30,6 @@ typedef struct s_plane
 	}						obj;
 }	t_plane;
 
-extern t_func	plane(enum e_element_func func);
-extern int		debug_plane(void *param);
 extern int		hit_plane(t_plane *self, t_ray *r, t_range ra, t_hit *h);
 extern int		init_plane(t_plane *self);
